ref(pageFilters): Consolidate multipleProjectSelector files
