+++ static/app/components/projectPageFilter.tsx
@@ -7,8 +7,8 @@ import partition from 'lodash/partition';
 import {updateProjects} from 'sentry/actionCreators/pageFilters';
 import GuideAnchor from 'sentry/components/assistant/guideAnchor';
 import Badge from 'sentry/components/badge';
-import MultipleProjectSelector from 'sentry/components/organizations/multipleProjectSelector';
 import PageFilterDropdownButton from 'sentry/components/organizations/pageFilters/pageFilterDropdownButton';
+import MultipleProjectSelector from 'sentry/components/organizations/projectSelector/multiple';
 import PlatformList from 'sentry/components/platformList';
 import {ALL_ACCESS_PROJECTS} from 'sentry/constants/pageFilters';
 import {IconProject} from 'sentry/icons';
